Preparing the game for a major redesign :)

+++ Source/World/Generators/Generation_Type.cpp
@@ -53,9 +53,6 @@ const Biome& World_Generation_Type::getBiome(int biomeKey)
     ///@TODO Return an "error biome" rather than crash the game
 }
 
-
-
-
 bool World_Generation_Type::parseLine(const std::string& line, std::ifstream& inFile)
 {
     if (areStringsSame(line, "Noise"))
@@ -75,11 +72,11 @@ bool World_Generation_Type::parseLine(const std::string& line, std::ifstream& in
 
 void World_Generation_Type::loadNoise(std::ifstream& inFile)
 {
-    int octaves,
-        amplitude,
-        heightOffset;
+    uint64_t    octaves,
+                amplitude,
+                heightOffset;
 
-    float   roughness,
+    double  roughness,
             smoothness;
